requirements.txt
discord.py==2.3.2
Requests==2.31.0
py-cord==2.4.1